+++ rust/operator-binary/src/crd/mod.rs
@@ -6,6 +6,10 @@ use serde::{Deserialize, Serialize};
 use shell_escape::escape;
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_operator::{
+    builder::pod::volume::{
+        ListenerOperatorVolumeSourceBuilder, ListenerOperatorVolumeSourceBuilderError,
+        ListenerReference, VolumeBuilder,
+    },
     commons::{
         affinity::StackableAffinity,
         cluster_operation::ClusterOperation,
@@ -21,10 +25,11 @@ use stackable_operator::{
     },
     k8s_openapi::{
         DeepMerge,
-        api::core::v1::{EnvVar, PodTemplateSpec},
+        api::core::v1::{EnvVar, PersistentVolumeClaim, PodTemplateSpec, Volume},
         apimachinery::pkg::api::resource::Quantity,
     },
     kube::{CustomResource, ResourceExt, runtime::reflector::ObjectRef},
+    kvp::Labels,
     product_config_utils::Configuration,
     product_logging::{self, spec::Logging},
     role_utils::{GenericRoleConfig, JavaCommonConfig, Role, RoleGroupRef},
@@ -59,8 +64,6 @@ pub const SSL_CLIENT_XML: &str = "ssl-client.xml";
 
 pub const HBASE_CLUSTER_DISTRIBUTED: &str = "hbase.cluster.distributed";
 pub const HBASE_ROOTDIR: &str = "hbase.rootdir";
-pub const HBASE_UNSAFE_REGIONSERVER_HOSTNAME_DISABLE_MASTER_REVERSEDNS: &str =
-    "hbase.unsafe.regionserver.hostname.disable.master.reversedns";
 
 pub const HBASE_UI_PORT_NAME_HTTP: &str = "ui-http";
 pub const HBASE_UI_PORT_NAME_HTTPS: &str = "ui-https";
@@ -76,9 +79,13 @@ pub const HBASE_REGIONSERVER_PORT: u16 = 16020;
 pub const HBASE_REGIONSERVER_UI_PORT: u16 = 16030;
 pub const HBASE_REST_PORT: u16 = 8080;
 pub const HBASE_REST_UI_PORT: u16 = 8085;
+pub const LISTENER_VOLUME_NAME: &str = "listener";
+pub const LISTENER_VOLUME_DIR: &str = "/stackable/listener";
 
 const DEFAULT_REGION_MOVER_TIMEOUT: Duration = Duration::from_minutes_unchecked(59);
 const DEFAULT_REGION_MOVER_DELTA_TO_SHUTDOWN: Duration = Duration::from_minutes_unchecked(1);
+
+const DEFAULT_LISTENER_CLASS: &str = "cluster-internal";
 
 #[derive(Snafu, Debug)]
 pub enum Error {
@@ -102,6 +109,27 @@ pub enum Error {
 
     #[snafu(display("incompatible merge types"))]
     IncompatibleMergeTypes,
+
+    #[snafu(display("empty values for role-group are not permitted"))]
+    EmptyRoleGroup,
+
+    #[snafu(display("role-group not found by name"))]
+    RoleGroupNotFound,
+
+    #[snafu(display("failed to build listener volume source"))]
+    ListenerVolumeSource {
+        source: ListenerOperatorVolumeSourceBuilderError,
+    },
+
+    #[snafu(display("failed to build listener volume"))]
+    BuildListenerVolume {
+        source: ListenerOperatorVolumeSourceBuilderError,
+    },
+
+    #[snafu(display("failed to build listener pvc"))]
+    BuildListenerPvc {
+        source: ListenerOperatorVolumeSourceBuilderError,
+    },
 }
 
 #[versioned(version(name = "v1alpha1"))]
@@ -171,18 +199,6 @@ pub mod versioned {
         /// for a ZooKeeper cluster.
         pub zookeeper_config_map_name: String,
 
-        /// This field controls which type of Service the Operator creates for this HbaseCluster:
-        ///
-        /// * cluster-internal: Use a ClusterIP service
-        ///
-        /// * external-unstable: Use a NodePort service
-        ///
-        /// This is a temporary solution with the goal to keep yaml manifests forward compatible.
-        /// In the future, this setting will control which [ListenerClass](DOCS_BASE_URL_PLACEHOLDER/listener-operator/listenerclass.html)
-        /// will be used to expose the service, and ListenerClass names will stay the same, allowing for a non-breaking change.
-        #[serde(default)]
-        pub listener_class: CurrentlySupportedListenerClasses,
-
         /// Settings related to user [authentication](DOCS_BASE_URL_PLACEHOLDER/usage-guide/security).
         pub authentication: Option<AuthenticationConfig>,
 
@@ -212,6 +228,11 @@ impl v1alpha1::HbaseCluster {
         let defaults =
             AnyConfigFragment::default_for(role, &self.name_any(), hdfs_discovery_cm_name);
 
+        // Trivial values for role-groups are not allowed
+        if role_group.is_empty() {
+            return Err(Error::EmptyRoleGroup);
+        }
+
         let (mut role_config, mut role_group_config) = match role {
             HbaseRole::RegionServer => {
                 let role = self
@@ -227,7 +248,7 @@ impl v1alpha1::HbaseCluster {
                     .role_groups
                     .get(role_group)
                     .map(|rg| rg.config.config.clone())
-                    .unwrap_or_default();
+                    .context(RoleGroupNotFoundSnafu)?;
 
                 (
                     AnyConfigFragment::RegionServer(role_config),
@@ -249,7 +270,7 @@ impl v1alpha1::HbaseCluster {
                     .role_groups
                     .get(role_group)
                     .map(|rg| rg.config.config.clone())
-                    .unwrap_or_default();
+                    .context(RoleGroupNotFoundSnafu)?;
 
                 // Retrieve role resource config
                 (
@@ -269,7 +290,7 @@ impl v1alpha1::HbaseCluster {
                     .role_groups
                     .get(role_group)
                     .map(|rg| rg.config.config.clone())
-                    .unwrap_or_default();
+                    .context(RoleGroupNotFoundSnafu)?;
 
                 // Retrieve role resource config
                 (
@@ -526,7 +547,7 @@ impl v1alpha1::HbaseCluster {
     }
 
     /// Name of the port used by the Web UI, which depends on HTTPS usage
-    fn ui_port_name(&self) -> String {
+    pub fn ui_port_name(&self) -> String {
         if self.has_https_enabled() {
             HBASE_UI_PORT_NAME_HTTPS
         } else {
@@ -550,27 +571,6 @@ pub fn merged_env(rolegroup_config: Option<&BTreeMap<String, String>>) -> Vec<En
         vec![]
     };
     merged_env
-}
-
-// TODO: Temporary solution until listener-operator is finished
-#[derive(Clone, Debug, Default, Display, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
-#[serde(rename_all = "PascalCase")]
-pub enum CurrentlySupportedListenerClasses {
-    #[default]
-    #[serde(rename = "cluster-internal")]
-    ClusterInternal,
-
-    #[serde(rename = "external-unstable")]
-    ExternalUnstable,
-}
-
-impl CurrentlySupportedListenerClasses {
-    pub fn k8s_service_type(&self) -> String {
-        match self {
-            CurrentlySupportedListenerClasses::ClusterInternal => "ClusterIP".to_string(),
-            CurrentlySupportedListenerClasses::ExternalUnstable => "NodePort".to_string(),
-        }
-    }
 }
 
 #[derive(Clone, Debug, Deserialize, Eq, Hash, JsonSchema, PartialEq, Serialize)]
@@ -696,6 +696,7 @@ impl HbaseRole {
             affinity: get_affinity(cluster_name, self, hdfs_discovery_cm_name),
             graceful_shutdown_timeout: Some(graceful_shutdown_timeout),
             requested_secret_lifetime: Some(requested_secret_lifetime),
+            listener_class: Some(DEFAULT_LISTENER_CLASS.to_string()),
         }
     }
 
@@ -707,6 +708,54 @@ impl HbaseRole {
             // instead of just letting the Display impl do it's thing ;P
             HbaseRole::RestServer => "rest".to_string(),
         }
+    }
+
+    pub fn listener_volume(
+        &self,
+        merged_config: &AnyServiceConfig,
+        recommended_labels: &Labels,
+    ) -> Result<Option<Volume>, Error> {
+        let volume = match &self {
+            // Master and regionservers should use ephemeral listener volumes
+            // since clients pull the latest address from ZooKeeper
+            HbaseRole::Master | HbaseRole::RegionServer => Some(
+                VolumeBuilder::new(LISTENER_VOLUME_NAME)
+                    .ephemeral(
+                        ListenerOperatorVolumeSourceBuilder::new(
+                            &ListenerReference::ListenerClass(
+                                merged_config.listener_class().to_string(),
+                            ),
+                            recommended_labels,
+                        )
+                        .context(ListenerVolumeSourceSnafu)?
+                        .build_ephemeral()
+                        .context(BuildListenerVolumeSnafu)?,
+                    )
+                    .build(),
+            ),
+            HbaseRole::RestServer => None,
+        };
+        Ok(volume)
+    }
+
+    pub fn listener_pvc(
+        &self,
+        merged_config: &AnyServiceConfig,
+        recommended_labels: &Labels,
+    ) -> Result<Option<Vec<PersistentVolumeClaim>>, Error> {
+        let pvc = match &self {
+            HbaseRole::Master | HbaseRole::RegionServer => None,
+            HbaseRole::RestServer => Some(vec![
+                ListenerOperatorVolumeSourceBuilder::new(
+                    &ListenerReference::ListenerClass(merged_config.listener_class().to_string()),
+                    recommended_labels,
+                )
+                .context(ListenerVolumeSourceSnafu)?
+                .build_pvc(LISTENER_VOLUME_NAME.to_string())
+                .context(BuildListenerPvcSnafu)?,
+            ]),
+        };
+        Ok(pvc)
     }
 }
 
@@ -796,6 +845,7 @@ impl AnyConfigFragment {
                         cli_opts: None,
                     },
                     requested_secret_lifetime: Some(HbaseRole::DEFAULT_REGION_SECRET_LIFETIME),
+                    listener_class: Some(DEFAULT_LISTENER_CLASS.to_string()),
                 })
             }
             HbaseRole::RestServer => AnyConfigFragment::RestServer(HbaseConfigFragment {
@@ -807,6 +857,7 @@ impl AnyConfigFragment {
                     HbaseRole::DEFAULT_REST_SERVER_GRACEFUL_SHUTDOWN_TIMEOUT,
                 ),
                 requested_secret_lifetime: Some(HbaseRole::DEFAULT_REST_SECRET_LIFETIME),
+                listener_class: Some(DEFAULT_LISTENER_CLASS.to_string()),
             }),
             HbaseRole::Master => AnyConfigFragment::Master(HbaseConfigFragment {
                 hbase_rootdir: None,
@@ -817,6 +868,7 @@ impl AnyConfigFragment {
                     HbaseRole::DEFAULT_MASTER_GRACEFUL_SHUTDOWN_TIMEOUT,
                 ),
                 requested_secret_lifetime: Some(HbaseRole::DEFAULT_MASTER_SECRET_LIFETIME),
+                listener_class: Some(DEFAULT_LISTENER_CLASS.to_string()),
             }),
         }
     }
@@ -894,6 +946,9 @@ pub struct HbaseConfig {
     /// Please note that this can be shortened by the `maxCertificateLifetime` setting on the SecretClass issuing the TLS certificate.
     #[fragment_attrs(serde(default))]
     pub requested_secret_lifetime: Option<Duration>,
+
+    /// This field controls which [ListenerClass](DOCS_BASE_URL_PLACEHOLDER/listener-operator/listenerclass.html) is used to expose this rolegroup.
+    pub listener_class: String,
 }
 
 impl Configuration for HbaseConfigFragment {
@@ -950,10 +1005,6 @@ impl Configuration for HbaseConfigFragment {
             HBASE_SITE_XML => {
                 result.insert(
                     HBASE_CLUSTER_DISTRIBUTED.to_string(),
-                    Some("true".to_string()),
-                );
-                result.insert(
-                    HBASE_UNSAFE_REGIONSERVER_HOSTNAME_DISABLE_MASTER_REVERSEDNS.to_string(),
                     Some("true".to_string()),
                 );
                 result.insert(HBASE_ROOTDIR.to_string(), self.hbase_rootdir.clone());
@@ -1047,6 +1098,9 @@ pub struct RegionServerConfig {
     /// The operator will compute a timeout period for the region move that will not exceed the graceful shutdown timeout.
     #[fragment_attrs(serde(default))]
     pub region_mover: RegionMover,
+
+    /// This field controls which [ListenerClass](DOCS_BASE_URL_PLACEHOLDER/listener-operator/listenerclass.html) is used to expose this rolegroup.
+    pub listener_class: String,
 }
 
 impl Configuration for RegionServerConfigFragment {
@@ -1101,10 +1155,6 @@ impl Configuration for RegionServerConfigFragment {
             HBASE_SITE_XML => {
                 result.insert(
                     HBASE_CLUSTER_DISTRIBUTED.to_string(),
-                    Some("true".to_string()),
-                );
-                result.insert(
-                    HBASE_UNSAFE_REGIONSERVER_HOSTNAME_DISABLE_MASTER_REVERSEDNS.to_string(),
                     Some("true".to_string()),
                 );
                 result.insert(HBASE_ROOTDIR.to_string(), self.hbase_rootdir.clone());
@@ -1169,6 +1219,14 @@ impl AnyServiceConfig {
             AnyServiceConfig::Master(config) => config.requested_secret_lifetime,
             AnyServiceConfig::RegionServer(config) => config.requested_secret_lifetime,
             AnyServiceConfig::RestServer(config) => config.requested_secret_lifetime,
+        }
+    }
+
+    pub fn listener_class(&self) -> String {
+        match self {
+            AnyServiceConfig::Master(config) => config.listener_class.clone(),
+            AnyServiceConfig::RegionServer(config) => config.listener_class.clone(),
+            AnyServiceConfig::RestServer(config) => config.listener_class.clone(),
         }
     }