+++ rust/operator-binary/src/hbase_controller.rs
@@ -8,6 +8,7 @@ use std::{
 };
 
 use const_format::concatcp;
+use indoc::formatdoc;
 use product_config::{
     ProductConfigManager,
     types::PropertyNameKind,
@@ -70,9 +71,11 @@ use crate::{
         construct_role_specific_non_heap_jvm_args,
     },
     crd::{
-        APP_NAME, AnyServiceConfig, Container, HBASE_ENV_SH, HBASE_REST_PORT_NAME_HTTP,
-        HBASE_REST_PORT_NAME_HTTPS, HBASE_SITE_XML, HbaseClusterStatus, HbaseRole,
-        JVM_SECURITY_PROPERTIES_FILE, SSL_CLIENT_XML, SSL_SERVER_XML, merged_env, v1alpha1,
+        APP_NAME, AnyServiceConfig, Container, HBASE_ENV_SH, HBASE_MASTER_PORT,
+        HBASE_MASTER_UI_PORT, HBASE_REGIONSERVER_PORT, HBASE_REGIONSERVER_UI_PORT,
+        HBASE_REST_PORT_NAME_HTTP, HBASE_REST_PORT_NAME_HTTPS, HBASE_SITE_XML, HbaseClusterStatus,
+        HbaseRole, JVM_SECURITY_PROPERTIES_FILE, LISTENER_VOLUME_DIR, LISTENER_VOLUME_NAME,
+        SSL_CLIENT_XML, SSL_SERVER_XML, merged_env, v1alpha1,
     },
     discovery::build_discovery_configmap,
     kerberos::{
@@ -101,6 +104,9 @@ const HBASE_CONFIG_TMP_DIR: &str = "/stackable/tmp/hbase";
 const HBASE_LOG_CONFIG_TMP_DIR: &str = "/stackable/tmp/log_config";
 
 const DOCKER_IMAGE_BASE_NAME: &str = "hbase";
+
+const HBASE_MASTER_PORT_NAME: &str = "master";
+const HBASE_REGIONSERVER_PORT_NAME: &str = "regionserver";
 
 pub struct Ctx {
     pub client: stackable_operator::client::Client,
@@ -131,9 +137,6 @@ pub enum Error {
     #[snafu(display("object defines no regionserver role"))]
     NoRegionServerRole,
 
-    #[snafu(display("failed to calculate global service name"))]
-    GlobalServiceNameNotFound,
-
     #[snafu(display("failed to create cluster resources"))]
     CreateClusterResources {
         source: stackable_operator::cluster_resources::Error,
@@ -141,11 +144,6 @@ pub enum Error {
 
     #[snafu(display("failed to delete orphaned resources"))]
     DeleteOrphanedResources {
-        source: stackable_operator::cluster_resources::Error,
-    },
-
-    #[snafu(display("failed to apply global Service"))]
-    ApplyRoleService {
         source: stackable_operator::cluster_resources::Error,
     },
 
@@ -312,6 +310,17 @@ pub enum Error {
 
     #[snafu(display("failed to construct JVM arguments"))]
     ConstructJvmArgument { source: crate::config::jvm::Error },
+
+    #[snafu(display("failed to build Labels"))]
+    LabelBuild {
+        source: stackable_operator::kvp::LabelError,
+    },
+
+    #[snafu(display("failed to build listener volume"))]
+    ListenerVolume { source: crate::crd::Error },
+
+    #[snafu(display("failed to build listener persistent volume claim"))]
+    ListenerPersistentVolumeClaim { source: crate::crd::Error },
 }
 
 type Result<T, E = Error> = std::result::Result<T, E>;
@@ -373,26 +382,6 @@ pub async fn reconcile_hbase(
     )
     .context(CreateClusterResourcesSnafu)?;
 
-    let region_server_role_service =
-        build_region_server_role_service(hbase, &resolved_product_image)?;
-    cluster_resources
-        .add(client, region_server_role_service)
-        .await
-        .context(ApplyRoleServiceSnafu)?;
-
-    // discovery config map
-    let discovery_cm = build_discovery_configmap(
-        hbase,
-        &client.kubernetes_cluster_info,
-        &zookeeper_connection_information,
-        &resolved_product_image,
-    )
-    .context(BuildDiscoveryConfigMapSnafu)?;
-    cluster_resources
-        .add(client, discovery_cm)
-        .await
-        .context(ApplyDiscoveryConfigMapSnafu)?;
-
     let (rbac_sa, rbac_rolebinding) = build_rbac_resources(
         hbase,
         APP_NAME,
@@ -441,7 +430,6 @@ pub async fn reconcile_hbase(
             )?;
             let rg_statefulset = build_rolegroup_statefulset(
                 hbase,
-                &client.kubernetes_cluster_info,
                 &hbase_role,
                 &rolegroup,
                 rolegroup_config,
@@ -482,6 +470,20 @@ pub async fn reconcile_hbase(
         }
     }
 
+    // Discovery CM will fail to build until the rest of the cluster has been deployed, so do it last
+    // so that failure won't inhibit the rest of the cluster from booting up.
+    let discovery_cm = build_discovery_configmap(
+        hbase,
+        &client.kubernetes_cluster_info,
+        &zookeeper_connection_information,
+        &resolved_product_image,
+    )
+    .context(BuildDiscoveryConfigMapSnafu)?;
+    cluster_resources
+        .add(client, discovery_cm)
+        .await
+        .context(ApplyDiscoveryConfigMapSnafu)?;
+
     let cluster_operation_cond_builder =
         ClusterOperationsConditionBuilder::new(&hbase.spec.cluster_operation);
 
@@ -499,59 +501,6 @@ pub async fn reconcile_hbase(
         .context(ApplyStatusSnafu)?;
 
     Ok(Action::await_change())
-}
-
-/// The server-role service is the primary endpoint that should be used by clients that do not perform internal load balancing,
-/// including targets outside of the cluster.
-pub fn build_region_server_role_service(
-    hbase: &v1alpha1::HbaseCluster,
-    resolved_product_image: &ResolvedProductImage,
-) -> Result<Service> {
-    let role = HbaseRole::RegionServer;
-    let role_name = role.to_string();
-    let role_svc_name = hbase
-        .server_role_service_name()
-        .context(GlobalServiceNameNotFoundSnafu)?;
-    let ports = hbase
-        .ports(&role)
-        .into_iter()
-        .map(|(name, value)| ServicePort {
-            name: Some(name),
-            port: i32::from(value),
-            protocol: Some("TCP".to_string()),
-            ..ServicePort::default()
-        })
-        .collect();
-
-    let metadata = ObjectMetaBuilder::new()
-        .name_and_namespace(hbase)
-        .name(&role_svc_name)
-        .ownerreference_from_resource(hbase, None, Some(true))
-        .context(ObjectMissingMetadataForOwnerRefSnafu)?
-        .with_recommended_labels(build_recommended_labels(
-            hbase,
-            &resolved_product_image.app_version_label,
-            &role_name,
-            "global",
-        ))
-        .context(ObjectMetaSnafu)?
-        .build();
-
-    let service_selector_labels =
-        Labels::role_selector(hbase, APP_NAME, &role_name).context(BuildLabelSnafu)?;
-
-    let service_spec = ServiceSpec {
-        type_: Some(hbase.spec.cluster_config.listener_class.k8s_service_type()),
-        ports: Some(ports),
-        selector: Some(service_selector_labels.into()),
-        ..ServiceSpec::default()
-    };
-
-    Ok(Service {
-        metadata,
-        spec: Some(service_spec),
-        status: None,
-    })
 }
 
 /// The rolegroup [`ConfigMap`] configures the rolegroup based on the configuration given by the administrator
@@ -587,6 +536,71 @@ fn build_rolegroup_config_map(
                 );
                 hbase_site_config
                     .extend(hbase_opa_config.map_or(vec![], |config| config.hbase_site_config()));
+
+                // Set flag to override default behaviour, which is that the
+                // RPC client should bind the client address (forcing outgoing
+                // RPC traffic to happen from the same network interface that
+                // the RPC server is bound on).
+                hbase_site_config.insert(
+                    "hbase.client.rpc.bind.address".to_string(),
+                    "false".to_string(),
+                );
+
+                match hbase_role {
+                    HbaseRole::Master => {
+                        hbase_site_config.insert(
+                            "hbase.master.ipc.address".to_string(),
+                            "0.0.0.0".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.master.ipc.port".to_string(),
+                            HBASE_MASTER_PORT.to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.master.hostname".to_string(),
+                            "${HBASE_SERVICE_HOST}".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.master.port".to_string(),
+                            "${HBASE_SERVICE_PORT}".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.master.info.port".to_string(),
+                            "${HBASE_INFO_PORT}".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.master.bound.info.port".to_string(),
+                            HBASE_MASTER_UI_PORT.to_string(),
+                        );
+                    }
+                    HbaseRole::RegionServer => {
+                        hbase_site_config.insert(
+                            "hbase.regionserver.ipc.address".to_string(),
+                            "0.0.0.0".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.regionserver.ipc.port".to_string(),
+                            HBASE_REGIONSERVER_PORT.to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.unsafe.regionserver.hostname".to_string(),
+                            "${HBASE_SERVICE_HOST}".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.regionserver.port".to_string(),
+                            "${HBASE_SERVICE_PORT}".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.regionserver.info.port".to_string(),
+                            "${HBASE_INFO_PORT}".to_string(),
+                        );
+                        hbase_site_config.insert(
+                            "hbase.regionserver.bound.info.port".to_string(),
+                            HBASE_REGIONSERVER_UI_PORT.to_string(),
+                        );
+                    }
+                    HbaseRole::RestServer => {}
+                };
 
                 // configOverride come last
                 hbase_site_config.extend(config.clone());
@@ -721,7 +735,7 @@ fn build_rolegroup_service(
 
     let metadata = ObjectMetaBuilder::new()
         .name_and_namespace(hbase)
-        .name(rolegroup.object_name())
+        .name(headless_service_name(&rolegroup.object_name()))
         .ownerreference_from_resource(hbase, None, Some(true))
         .context(ObjectMissingMetadataForOwnerRefSnafu)?
         .with_recommended_labels(build_recommended_labels(
@@ -761,7 +775,6 @@ fn build_rolegroup_service(
 #[allow(clippy::too_many_arguments)]
 fn build_rolegroup_statefulset(
     hbase: &v1alpha1::HbaseCluster,
-    cluster_info: &KubernetesClusterInfo,
     hbase_role: &HbaseRole,
     rolegroup_ref: &RoleGroupRef<v1alpha1::HbaseCluster>,
     rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
@@ -785,14 +798,14 @@ fn build_rolegroup_statefulset(
     let probe_template = match hbase_role {
         HbaseRole::Master => Probe {
             tcp_socket: Some(TCPSocketAction {
-                port: IntOrString::String("master".to_string()),
+                port: IntOrString::String(HBASE_MASTER_PORT_NAME.to_string()),
                 ..TCPSocketAction::default()
             }),
             ..Probe::default()
         },
         HbaseRole::RegionServer => Probe {
             tcp_socket: Some(TCPSocketAction {
-                port: IntOrString::String("regionserver".to_string()),
+                port: IntOrString::String(HBASE_REGIONSERVER_PORT_NAME.to_string()),
                 ..TCPSocketAction::default()
             }),
             ..Probe::default()
@@ -856,15 +869,30 @@ fn build_rolegroup_statefulset(
         },
     ]);
 
+    let role_name = hbase_role.cli_role_name();
     let mut hbase_container = ContainerBuilder::new("hbase").expect("ContainerBuilder not created");
+
+    let rest_http_port_name = if hbase.has_https_enabled() {
+        HBASE_REST_PORT_NAME_HTTPS
+    } else {
+        HBASE_REST_PORT_NAME_HTTP
+    };
+
     hbase_container
         .image_from_product_image(resolved_product_image)
-        .command(vec!["/stackable/hbase/bin/hbase-entrypoint.sh".to_string()])
-        .args(vec![
-            hbase_role.cli_role_name(),
-            hbase_service_domain_name(hbase, rolegroup_ref, cluster_info)?,
-            hbase.service_port(hbase_role).to_string(),
-        ])
+        .command(command())
+        .args(vec![formatdoc! {"
+            {entrypoint} {role} {port} {port_name} {ui_port_name}",
+            entrypoint = "/stackable/hbase/bin/hbase-entrypoint.sh".to_string(),
+            role = role_name,
+            port = hbase.service_port(hbase_role).to_string(),
+            port_name = match hbase_role {
+                HbaseRole::Master => HBASE_MASTER_PORT_NAME,
+                HbaseRole::RegionServer => HBASE_REGIONSERVER_PORT_NAME,
+                HbaseRole::RestServer => rest_http_port_name,
+            },
+            ui_port_name = hbase.ui_port_name(),
+        }])
         .add_env_vars(merged_env)
         // Needed for the `containerdebug` process to log it's tracing information to.
         .add_env_var(
@@ -879,6 +907,8 @@ fn build_rolegroup_statefulset(
         .context(AddVolumeMountSnafu)?
         .add_volume_mount("log", STACKABLE_LOG_DIR)
         .context(AddVolumeMountSnafu)?
+        .add_volume_mount(LISTENER_VOLUME_NAME, LISTENER_VOLUME_DIR)
+        .context(AddVolumeMountSnafu)?
         .add_container_ports(ports)
         .resources(merged_config.resources().clone().into())
         .startup_probe(startup_probe)
@@ -887,13 +917,17 @@ fn build_rolegroup_statefulset(
 
     let mut pod_builder = PodBuilder::new();
 
+    let recommended_object_labels = build_recommended_labels(
+        hbase,
+        hbase_version,
+        &rolegroup_ref.role,
+        &rolegroup_ref.role_group,
+    );
+    let recommended_labels =
+        Labels::recommended(recommended_object_labels.clone()).context(LabelBuildSnafu)?;
+
     let pb_metadata = ObjectMetaBuilder::new()
-        .with_recommended_labels(build_recommended_labels(
-            hbase,
-            hbase_version,
-            &rolegroup_ref.role,
-            &rolegroup_ref.role_group,
-        ))
+        .with_recommended_labels(recommended_object_labels)
         .context(ObjectMetaSnafu)?
         .build();
 
@@ -999,7 +1033,21 @@ fn build_rolegroup_statefulset(
         }
     }
 
+    let listener_pvc = hbase_role
+        .listener_pvc(merged_config, &recommended_labels)
+        .context(ListenerPersistentVolumeClaimSnafu)?;
+
+    if let Some(listener_volume) = hbase_role
+        .listener_volume(merged_config, &recommended_labels)
+        .context(ListenerVolumeSnafu)?
+    {
+        pod_builder
+            .add_volume(listener_volume)
+            .context(AddVolumeSnafu)?;
+    };
+
     let mut pod_template = pod_builder.build_template();
+
     hbase.merge_pod_overrides(&mut pod_template, hbase_role, rolegroup_ref);
 
     let metadata = ObjectMetaBuilder::new()
@@ -1031,8 +1079,9 @@ fn build_rolegroup_statefulset(
             match_labels: Some(statefulset_match_labels.into()),
             ..LabelSelector::default()
         },
-        service_name: Some(rolegroup_ref.object_name()),
+        service_name: Some(headless_service_name(&rolegroup_ref.object_name())),
         template: pod_template,
+        volume_claim_templates: listener_pvc,
         ..StatefulSetSpec::default()
     };
 
@@ -1041,6 +1090,17 @@ fn build_rolegroup_statefulset(
         spec: Some(statefulset_spec),
         status: None,
     })
+}
+
+/// Returns the container command.
+fn command() -> Vec<String> {
+    vec![
+        "/bin/bash".to_string(),
+        "-x".to_string(),
+        "-euo".to_string(),
+        "pipefail".to_string(),
+        "-c".to_string(),
+    ]
 }
 
 fn write_hbase_env_sh<'a, T>(properties: T) -> String
@@ -1104,6 +1164,7 @@ fn build_hbase_env_sh(
     let role_specific_non_heap_jvm_args =
         construct_role_specific_non_heap_jvm_args(hbase, hbase_role, role_group)
             .context(ConstructJvmArgumentSnafu)?;
+
     match hbase_role {
         HbaseRole::Master => {
             result.insert(
@@ -1128,26 +1189,8 @@ fn build_hbase_env_sh(
     Ok(result)
 }
 
-/// Build the domain name of an HBase service pod.
-/// The hbase-entrypoint.sh script uses this to build the fully qualified name of a pod
-/// by appending it to the `HOSTNAME` environment variable.
-/// This name is required by the RegionMover to function properly.
-fn hbase_service_domain_name(
-    hbase: &v1alpha1::HbaseCluster,
-    rolegroup_ref: &RoleGroupRef<v1alpha1::HbaseCluster>,
-    cluster_info: &KubernetesClusterInfo,
-) -> Result<String, Error> {
-    let hbase_cluster_name = rolegroup_ref.object_name();
-    let pod_namespace = hbase
-        .metadata
-        .namespace
-        .clone()
-        .context(ObjectHasNoNamespaceSnafu)?;
-    let cluster_domain = &cluster_info.cluster_domain;
-
-    Ok(format!(
-        "{hbase_cluster_name}.{pod_namespace}.svc.{cluster_domain}"
-    ))
+fn headless_service_name(role_group_name: &str) -> String {
+    format!("{name}-headless", name = role_group_name)
 }
 
 #[cfg(test)]